Updated syntax

+++ src/ssl/context.cr
@@ -49,7 +49,8 @@ module OpenSSL::SSL
   class Context
     alias VerifyCallback = (Bool, X509::StoreContext) -> Bool
 
-    @@index : Int64
+    @@index : Int32
+
     @@index = begin
                 index = LibSSL.ssl_ctx_get_ex_new_index(0_i64, nil, nil, nil, nil)
                 if index < 0
